RAG/rag_requirements.txt
python-dotenv
langchainhub
langchain_huggingface
langchain
sentence_transformers
faiss-cpu
langchain-community
langsmith